==> Codes/Step_4_Filters.ipynb <==
<p style="text-align:center; font-size:50px; font-weight:bold; color: greenyellow; border: 10px solid red; border-radius: 20px; background-color: black; padding: 30px 0px;">Hampel + Savitzky Golay Filter</p>

<p style="text-align:center; font-size:50px; font-weight:bold; color: blue; border-radius: 20px; padding: 30px 0px;">Use Only If Data Is Too Much Noisy</p>

import pandas as pd
import numpy as np
from scipy.signal import savgol_filter

def hampel_filter(input_series, window_size=15, n_sigmas=3):
    new_series = input_series.copy()
    L = 1.4826 
    
    for i in range(window_size, len(input_series) - window_size):
        window_data = input_series[i - window_size : i + window_size]
        median = np.median(window_data)
        mad = L * np.median(np.abs(window_data - median))
        if mad == 0:
            continue
        threshold = n_sigmas * mad
        if np.abs(input_series[i] - median) > threshold:
            new_series[i] = median
    return new_series

df = pd.read_csv("csi_data_amp_phase.csv")

csi_cols = [col for col in df.columns if "_amp" in col or "_phase" in col]

for col in csi_cols:
    df[col] = hampel_filter(df[col].values, window_size=15, n_sigmas=3)
    df[col] = savgol_filter(df[col].values, window_length=5, polyorder=3)

df.to_csv("csi_dataset_finals.csv", index=False)
print("Final dataset with Hampel + Savitzky Golay filter saved as 'csi_dataset_finals.csv'")
